+++ misc/plugutils.c
@@ -28,11 +28,12 @@
 #include "network.h" // for OPENVAS_ENCAPS_IP
 
 #include <errno.h>               // for errno
+#include <gvm/base/cvss.h>       // for get_cvss_score_from_base_metrics
 #include <gvm/base/hosts.h>      // for g_vhost_t
 #include <gvm/base/networking.h> // for port_protocol_t
 #include <gvm/base/prefs.h>      // for prefs_get_bool
-#include <gvm/util/mqtt.h>       // for mqtt_reset
-#include <gvm/util/nvticache.h>  // for nvticache_initialized
+#include <gvm/util/mqtt.h>      // for mqtt_reset
+#include <gvm/util/nvticache.h> // for nvticache_initialized
 #include <json-glib/json-glib.h>
 #include <stdio.h>    // for snprintf
 #include <stdlib.h>   // for exit
@@ -406,6 +407,72 @@ make_table_driven_lsc_info_json_str (const char *scan_id, const char *ip_str,
 }
 
 /**
+ * @brief Get the nvti's severity vector and return the score as string.
+ *
+ * @param[in] nvti     The nvti where to get the severity vector from.
+ *
+ * @return string representing a float value, with max 1 decimal. Eg. -1, 3.2.
+ * NULL otherwise. Must be free()'d by the caller.
+ */
+static gchar *
+get_severity_score_from_vt (nvti_t *nvti)
+{
+  gchar *vector;
+  gchar *score_str;
+  double score;
+
+  vector = nvti_severity_vector_from_tag (nvti);
+  if (!vector)
+    return NULL;
+
+  score = get_cvss_score_from_base_metrics (vector);
+  g_free (vector);
+
+  score_str = g_strdup_printf ("%.1f", score);
+
+  return score_str;
+}
+
+/**
+ * @brief Build a json representation of the part related to nvti information
+ * of a result.
+ *
+ * OpenVAS result contains also the nvti name, QoD and score, which need to be
+ * taken from redis and the score must be calculated from the vector
+ *
+ * @param[in]  oid     The oid of the NVT.
+ * @param[out] builder Json builder to add data to.
+ *
+ */
+static void
+add_nvti_info_into_json_builder (const char *oid, JsonBuilder *builder)
+{
+  nvti_t *nvti;
+  gchar *score_str;
+
+  nvti = nvticache_get_nvt (oid);
+
+  if (!nvti)
+    {
+      g_warning ("%s: Plugin '%s' missing from nvticache.", __func__, oid);
+      return;
+    }
+
+  json_builder_set_member_name (builder, "name");
+  builder = json_builder_add_string_value (builder, nvti_name (nvti));
+
+  json_builder_set_member_name (builder, "qod");
+  builder = json_builder_add_string_value (builder, nvti_qod (nvti));
+
+  score_str = get_severity_score_from_vt (nvti);
+  json_builder_set_member_name (builder, "severity");
+  builder = json_builder_add_string_value (builder, score_str);
+  g_free (score_str);
+
+  nvti_free (nvti);
+}
+
+/**
  * @brief Build a json representation of a result.
  *
  * JSON result consists of scan_id, message type, host ip,  hostname, port
@@ -457,6 +524,9 @@ make_result_json_str (const gchar *scan_id, const gchar *type,
 
   json_builder_set_member_name (builder, "OID");
   json_builder_add_string_value (builder, oid);
+
+  if (oid != NULL && oid[0] != '\0')
+    add_nvti_info_into_json_builder (oid, builder);
 
   json_builder_set_member_name (builder, "value");
   json_builder_add_string_value (builder, action_str);